backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
bcrypt==4.0.1
